// src/api/estrutura/models/ChatConversationForkDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ChatConversationForkDto = {
  /**
   * @type string
   */
  sessionId: string;
  /**
   * @type string
   */
  userId: string;
  /**
   * @type string
   */
  contextType: string;
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string
   */
  model: string;
  /**
   * @type string
   */
  conversationId: string;
  /**
   * @type object
   */
  shareId?: object | null;
  /**
   * @type object
   */
  parentConversationId?: object | null;
  /**
   * @type object
   */
  rootConversationId?: object | null;
  /**
   * @type string
   */
  forkedFromConversationId: string;
};