// src/api/estrutura/models/ResEstruturaDependentesDTO.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ResEstruturaItemDTO } from './ResEstruturaItemDTO';

export type ResEstruturaDependentesDTO = {
  /**
   * @type object
   */
  target: ResEstruturaItemDTO;
  /**
   * @type array
   */
  depedencias: ResEstruturaItemDTO[];
};